+++ app/(main)/components/home-client-component.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useEffect, useState, useRef, Suspense } from "react"
+import { useState, useRef } from "react"
 import Image from "next/image"
 import Link from "next/link"
 import { motion, useInView } from "framer-motion"
@@ -14,29 +14,7 @@ import Autoplay from "embla-carousel-autoplay"
 import { HeroSection } from "@/components/hero-section"
 import { Skeleton } from "@/components/ui/skeleton"
 import { getValidImageUrl, handleImageError } from "@/lib/utils/image-utils"
-
-// Simple CountUp component remains the same
-function CountUp({ target, duration = 1.5 }: { target: number; duration?: number }) {
-  const [count, setCount] = useState(0)
-  useEffect(() => {
-    setCount(0)
-    const totalFrames = Math.min(duration / 16, 100) 
-    const increment = target > 0 ? target / totalFrames : 0
-    let currentCount = 0
-    let frame = 0
-    const counter = setInterval(() => {
-      currentCount = Math.min(currentCount + increment, target)
-      setCount(Math.floor(currentCount))
-      frame++
-      if (frame >= totalFrames) {
-        clearInterval(counter)
-        setCount(target)
-      }
-    }, 16)
-    return () => clearInterval(counter)
-  }, [target, duration])
-  return <>{count}</>
-}
+import { CountUp } from "@/components/ui/count-up"
 
 // Client Component Part
 interface HomeClientComponentProps {
